[Feat] [SCRUM-163] user patch me 연결 body 실패
