+++ src/service/nestjs/src/api/user/dto/request/index.ts
@@ -1,1 +1,14 @@
-export * from './update'
+import { ApiProperty, PartialType } from '@nestjs/swagger';
+
+export class Update {
+	@ApiProperty({ description: 'Nickname', example: 'younhwan' })
+	nickname: string;
+
+	@ApiProperty({ description: 'Profile image URI', nullable: true })
+	profileImageURI: string;
+	
+	@ApiProperty({ description: 'email', nullable: true })
+	email2fa: string;
+}
+
+export class meUpdate extends PartialType(Update) {}
